# Replace map with matrix in vari_value