## Summary
Replaced map with matrix in vari_vase. Separate templates for value and adjoint made adding block very simple.

@SteveBronder You did sme benchmarking when adding vari_value. Can you also check how this compares with your version with maps?