@SteveBronder 2020-07-02
Nice! I was worried overusing `chainable_alloc` would cause some slowdowns, let me put this into that branch I did for the performance testing