@SteveBronder 2020-07-02
Cool so it looks like they are both about the same speed up to the margin of error from the sun's gamma rays flipping bits on my computer. Being able to have simple block ops etc. feels v worth it and if it turns out to be not good we can always try to work out a way to go back to maps

Docs need fixed up, I'll give this a review!

#### This Branch

```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
stat_mul_add/2           458 ns          458 ns      1517299
stat_mul_add/4           794 ns          794 ns       882689
stat_mul_add/8          1699 ns         1699 ns       406741
stat_mul_add/16         5856 ns         5854 ns       121152
stat_mul_add/32        23806 ns        23803 ns        29446
stat_mul_add/64       117339 ns       117338 ns         5868
stat_mul_add/128      657300 ns       657295 ns         1023
stat_mul_add/256     4214419 ns      4214352 ns          166
stat_mul_add/512    33209180 ns     33202847 ns           21
stat_mul_add/1024  259836242 ns    259803267 ns            3
stat_mul_add/2048 1870464718 ns   1870105019 ns            1
```

#### Original Branch

```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
stat_mul_add/2           425 ns          425 ns      1660515
stat_mul_add/4           657 ns          656 ns      1082579
stat_mul_add/8          1609 ns         1608 ns       430214
stat_mul_add/16         5729 ns         5728 ns       116832
stat_mul_add/32        24107 ns        24102 ns        29017
stat_mul_add/64       119877 ns       119849 ns         5719
stat_mul_add/128      665804 ns       665748 ns         1007
stat_mul_add/256     4331115 ns      4330477 ns          161
stat_mul_add/512    34315033 ns     34303303 ns           21
stat_mul_add/1024  251633171 ns    251581429 ns            3
stat_mul_add/2048 1911467093 ns   1910883203 ns            1
```

Side note: I almost had a heart attack when running the benchmarks at first because static and dynamic matrices were the same speed, but I was compiling with `-O0 and -g`. Switching to O3 showed the gainz from before