@t4c1 2020-07-03
Nice, the benchmark shows this won't be significantly slower. But I am sure there are some cases where this version might be significantly faster (by avoiding copys with perfect forwarding).