@t4c1 2020-07-03
Btw this also allows us to add other views into the matrix in the same fashion as block. For example transpose and triangularView.