@SteveBronder 2020-07-03
Actually @t4c1 I think my benchmarks from before weren't capturing the right thing. The real cost we want to test here is the destruction so I wrote this benchmark which just creates a bunch of matrices and then calls `recover_memory()` 

https://gist.github.com/SteveBronder/47d5e5e7c6642018835bc01d21e95f29

#### This PR

```cpp
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
toss_me                        4.68 ns         4.68 ns    149975344
stat_create_destory/2           289 ns          289 ns      2438544
stat_create_destory/4           473 ns          472 ns      1484629
stat_create_destory/8           544 ns          544 ns      1254389
stat_create_destory/16         1045 ns         1044 ns       668203
stat_create_destory/32         2640 ns         2639 ns       265258
stat_create_destory/64        63937 ns        63925 ns        10909
stat_create_destory/128      326796 ns       326547 ns         2179
stat_create_destory/256     1859036 ns      1857238 ns          376
stat_create_destory/512     9530080 ns      9520333 ns           73
stat_create_destory/1024   40667242 ns     40590689 ns           17
stat_create_destory/2048  156111987 ns    155813226 ns            5
```

#### feature/vari-base-eigen

```cpp
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
toss_me                        4.69 ns         4.69 ns    149474006
stat_create_destory/2           129 ns          129 ns      5252528
stat_create_destory/4           152 ns          152 ns      4583588
stat_create_destory/8           232 ns          232 ns      3024206
stat_create_destory/16          734 ns          734 ns       973697
stat_create_destory/32         2367 ns         2366 ns       293801
stat_create_destory/64         9014 ns         9012 ns        77628
stat_create_destory/128       36883 ns        36879 ns        18747
stat_create_destory/256      550235 ns       549952 ns         1247
stat_create_destory/512     4353909 ns      4352943 ns          160
stat_create_destory/1024   20720279 ns     20711722 ns           34
stat_create_destory/2048   82233619 ns     82219982 ns            8
```

#### Diff 1 - (new / old) 

Negative means slower by X magnitude

size | time
-- | --
2 | -1.24031007751938
4 | -2.11184210526316
8 | -1.3448275862069
16 | -0.423705722070845
32 | -0.115335868187579
64 | -6.09307743510095
128 | -7.86034216305615
256 | -2.37862186156824
512 | -1.18885603718406
1024 | -0.962678301773832
2048 | -0.898396165684986

So in terms of destroying stuff it looks like this is a lot more expensive. Does that benchmark seem like it's measuring the right stuff to you? If so the question then is how much this cost carries relative to the niceties this gives us. I'm half and half on it. The other Q here is how expensive this will be relative to whatever the `block()` implementation will look like when using a map as the base