@SteveBronder 2020-07-03
And also in the largest case it's like `0.07` seconds. If we do 1000 gradient evaluations and recover_memory() after each then that's an extra 70 seconds which feels like it matters