@SteveBronder 2020-07-03
So I have this tho I'm not sure if I like it

https://godbolt.org/z/zpiKox

So the good is we keep using map and the stack, the bad is that we need to do another alloc so would add a new vari to the stack. And we have to make `val_` not const or a bunch of crazy compiler errors happen that I had a hard time sorting out. 

There should be a way to have a view/slice without doing an allocation tho. Something to do with the stride or something