@SteveBronder 2020-07-04
Okay I got rid of the alloc, still a little janky but I can fix it up this weekend

https://godbolt.org/z/xgXLmV