@SteveBronder 2020-07-04
Alright my b for so many messages but how about this?

https://godbolt.org/z/Rmeo-X

That keeps constness around and still let's us have just a view into memory while still allocating on the stack