@t4c1 2020-07-06
Those suggestions use new, but I don't see delete anywhere. So they are leaking memory? Also I have no idea what changed between different suggestions. Can you put them somewhere on github so I can see the diff?