@t4c1 2020-07-06
Your benchmark shows worst case performance for matrix. I also benchmarked best case (by always constructing vari_value from rvalue) and it is still slower than map. So I am closing this.