@bob-carpenter 2020-07-06
Yes, putting things on the stack is slower.  The stack has to be iterated and constructors have to be called.  But then it saves creating Map instances, which have their own construction cost.  

Is the slowdown for small sizes systematic? 

Is this all moot now that the issue's closed?